Add files via upload
The Dog Breeds app is an informative and interactive Android application designed for dog enthusiasts and pet owners. The app provides a comprehensive list of popular dog breeds, complete with pictures and brief descriptions. A key feature of the app is the ability to click on a breed's name or image, which will take users directly to a webpage with more information about the selected breed.